[SofaVolumetricData] Cleaning in DistanceGrid

+++ applications/plugins/SceneCreator/SceneCreator_test/SceneCreator_test.cpp
@@ -294,7 +294,7 @@ bool SceneCreator_test::createSphereSuccess()
     // Check MecaObj
     std::vector<MechanicalObject3*> mecaObjs;
     node->get<MechanicalObject3>(&mecaObjs, sofa::core::objectmodel::BaseContext::SearchDown);
-    EXPECT_EQ(mecaObjs.size(), 1);
+    EXPECT_EQ(mecaObjs.size(), (size_t)1);
 
 
     // check Grid
